requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.105.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mrz==0.6.2
numpy==1.24.4
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.2
Pillow==10.1.0
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pytesseract==0.3.10
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
readmrz==0.0.2
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.9.0
ujson==5.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
